+++ src/fields/formfields/Phone.php
@@ -207,6 +207,12 @@ class Phone extends FormField implements PreviewableFieldInterface
     public function validatePhone(ElementInterface $element)
     {
         $value = $element->getFieldValue($this->handle);
+
+        // Don't run validation if a field is not required and has no value for the phone number
+        if (!$this->required && empty($value->phone)) {
+            return;
+        }
+
         $isValid = SproutBaseFields::$app->phoneField->validate($value);
 
         if (!$isValid) {