+++ src/integrations/sproutemail/events/notificationevents/SaveEntryEvent.php
@@ -15,6 +15,7 @@ use Craft;
 use Twig\Error\LoaderError;
 use Twig\Error\RuntimeError;
 use Twig\Error\SyntaxError;
+use yii\base\Exception;
 
 
 /**
@@ -73,14 +74,13 @@ class SaveEntryEvent extends NotificationEvent
 
 
     /**
-     * @inheritdoc
-     *
      * @param array $context
      *
      * @return string
      * @throws LoaderError
      * @throws RuntimeError
      * @throws SyntaxError
+     * @throws Exception
      */
     public function getSettingsHtml($context = []): string
     {
@@ -138,24 +138,26 @@ class SaveEntryEvent extends NotificationEvent
         return null;
     }
 
-    public function rules(): array
+    /**
+     * @return array
+     */
+    public function defineRules(): array
     {
-        $rules = parent::rules();
+        $rules = parent::defineRules();
 
         $rules[] = [
             'whenNew', 'required', 'when' => function() {
                 return $this->whenUpdated == false;
             }
         ];
-
         $rules[] = [
             'whenUpdated', 'required', 'when' => function() {
                 return $this->whenNew == false;
             }
         ];
-
         $rules[] = [['whenNew', 'whenUpdated'], 'validateWhenTriggers'];
         $rules[] = [['event'], 'validateEvent'];
+        $rules[] = [['event'], 'validateCaptchas'];
         $rules[] = [['formIds'], 'validateFormIds'];
 
         return $rules;
@@ -199,6 +201,15 @@ class SaveEntryEvent extends NotificationEvent
 
         if (get_class($event->entry) !== Entry::class) {
             $this->addError('event', Craft::t('sprout-forms', 'Event Element does not match barrelstrength\sproutforms\elements\Entry class.'));
+        }
+    }
+
+    public function validateCaptchas()
+    {
+        $entry = $this->event->entry;
+
+        if ($entry->hasCaptchaErrors()) {
+            $this->addError('event', Craft::t('sprout-forms', 'Submitted entry has captcha errors.'));
         }
     }