+++ src/records/EntryStatus.php
@@ -3,7 +3,6 @@
 namespace barrelstrength\sproutforms\records;
 
 use craft\db\ActiveRecord;
-use craft\helpers\UrlHelper;
 
 /**
  * Class EntryStatus record
@@ -26,21 +25,4 @@ class EntryStatus extends ActiveRecord
     {
         return '{{%sproutforms_entrystatuses}}';
     }
-
-    /**
-     * @return string
-     */
-    public function getCpEditUrl(): string
-    {
-        return UrlHelper::cpUrl('sprout-forms/settings/orders-statuses/'.$this->id);
-    }
-
-    /**
-     * @return string
-     */
-    public function htmlLabel(): string
-    {
-        return '<span class="sproutFormsStatusLabel"><span class="status '.$this->color.'"></span> '.$this->name.'</span>';
-    }
-
 }