+++ src/rules/conditions/IsCondition.php
@@ -20,17 +20,7 @@ class IsCondition extends Condition
     }
 
     /**
-     * @return array
-     */
-    public function rules(): array
-    {
-        return [
-            [['inputValue'], 'validateCondition', 'skipOnEmpty' => false]
-        ];
-    }
-
-    /**
-     * @inheritDoc
+     * @return bool
      */
     public function validateCondition()
     {