+++ src/rules/conditions/IsNotCheckedCondition.php
@@ -20,16 +20,6 @@ class IsNotCheckedCondition extends Condition
     }
 
     /**
-     * @return array
-     */
-    public function rules(): array
-    {
-        return [
-            [['inputValue'], 'validateCondition', 'skipOnEmpty' => false]
-        ];
-    }
-
-    /**
      * @inheritDoc
      */
     public function validateCondition()