+++ CHANGELOG.md
@@ -19,6 +19,58 @@
 ### Removed
 - Removed `barrelstrength/sprout-base-lists` requirement (use Mailing List Reports)
 
+## 3.7.1.1 - 2020-01-18
+
+### Fixed 
+- Fixed composer release number 
+
+## 3.7.1 - 2020-01-18
+
+### Updated
+- Updated `barrelstrength/sprout-base-fields` to v1.2.3
+
+### Fixed 
+- Fixed array offset error in PHP 7.4 ([#405][405-sproutbasefields]) 
+
+[405-sproutbasefields]: https://github.com/barrelstrength/craft-sprout-forms/issues/405
+
+## 3.7.0 - 2020-01-17
+
+### Added
+- Added `barrelstrength\sproutforms\base\Captcha::$form`
+
+## 3.6.10 - 2020-01-16
+
+### Updated
+- Updated `barrelstrength/sprout-base-fields` to v1.2.2
+
+### Fixed 
+- Fixed error in address table migration
+
+## 3.6.9 - 2020-01-15
+
+### Added
+- Added `barrelstrength\sproutforms\models\EntryStatus::getCpEditUrl()`
+- Added `barrelstrength\sproutforms\models\EntryStatus::htmlLabel()`
+
+### Changed
+- Updated 'When a form entry is saved' Notification Event to only send notifications if captchas pass validation ([#396])
+- Updated `barrelstrength\sproutforms\models\EntryStatus::$isDefault` default value to `false`
+
+### Fixed
+- Fixed bug where optional Phone field would not validate with blank value ([#403])
+- Fixed bug where 'When a form is saved' Notification Event may not get migrated properly ([#400])
+- Fixed bug where custom Entry Status could not be deleted ([#368])
+
+[#368]: https://github.com/barrelstrength/craft-sprout-forms/issues/386
+[#396]: https://github.com/barrelstrength/craft-sprout-forms/issues/396
+[#400]: https://github.com/barrelstrength/craft-sprout-forms/issues/400
+[#403]: https://github.com/barrelstrength/craft-sprout-forms/issues/403
+
+### Removed
+- Removed `barrelstrength\sproutforms\records\EntryStatus::getCpEditUrl()`
+- Removed `barrelstrength\sproutforms\records\EntryStatus::htmlLabel()`
+
 ## 3.6.8 - 2020-01-09
 
 ### Updated